chore: Remove old SignsUI based suppression logic.
