+++ config/config.exs
@@ -40,9 +40,6 @@ config :concentrate,
       trip_updates: "https://cdn.mbta.com/realtime/TripUpdates.pb"
     ]
   ],
-  signs_stops_config: [
-    url: "s3://mbta-signs-dev/stops-config.json"
-  ],
   alerts: [
     url: "https://cdn.mbta.com/realtime/Alerts.pb"
   ],
@@ -78,8 +75,7 @@ config :concentrate,
     Concentrate.GroupFilter.SkippedStopOnAddedTrip,
     Concentrate.GroupFilter.TripDescriptorTimestamp,
     Concentrate.GroupFilter.UncertaintyValue,
-    Concentrate.GroupFilter.SuppressStopTimeUpdate,
-    Concentrate.GroupFilter.ScreenplaySuppressStopTimeUpdate
+    Concentrate.GroupFilter.SuppressStopTimeUpdate
   ],
   source_reporters: [
     Concentrate.SourceReporter.Basic,
@@ -118,8 +114,7 @@ config :concentrate,
     "http" => Concentrate.Producer.HTTPoison,
     "mqtt" => Concentrate.Producer.Mqtt,
     "mqtts" => Concentrate.Producer.Mqtt,
-    "mqtt+ssl" => Concentrate.Producer.Mqtt,
-    "s3" => Concentrate.Producer.S3
+    "mqtt+ssl" => Concentrate.Producer.Mqtt
   }
 
 import_config "#{config_env()}.exs"
