+++ lib/concentrate/filter/suppress/stop_prediction_status.ex
@@ -1,4 +1,4 @@
-defmodule Concentrate.Filter.Suppress.Screenplay.StopPredictionStatus do
+defmodule Concentrate.Filter.Suppress.StopPredictionStatus do
   @moduledoc """
   Server which stores a set of suppression_type, route_id, direction_id, and stop_ids which is used to
   filter out StopTimeUpdate structs for that combination.
