+++ test/concentrate/filter/suppress/stop_prediction_status_test.exs
@@ -1,7 +1,7 @@
-defmodule Concentrate.Filter.Suppress.Screenplay.StopPredictionStatusTest do
+defmodule Concentrate.Filter.Suppress.StopPredictionStatusTest do
   @moduledoc false
   use ExUnit.Case
-  import Concentrate.Filter.Suppress.Screenplay.StopPredictionStatus
+  import Concentrate.Filter.Suppress.StopPredictionStatus
   import ExUnit.CaptureLog
 
   @entries [
@@ -92,7 +92,7 @@ defmodule Concentrate.Filter.Suppress.Screenplay.StopPredictionStatusTest do
   end
 
   defp supervised(_) do
-    start_supervised(Concentrate.Filter.Suppress.Screenplay.StopPredictionStatus)
+    start_supervised(Concentrate.Filter.Suppress.StopPredictionStatus)
     :ok
   end
 end
