+++ test/concentrate/supervisor_test.exs
@@ -15,11 +15,11 @@ defmodule Concentrate.SupervisorTest do
 
   describe "children/1" do
     test "builds the right number of children" do
-      # currently, the right number is 6: HTTP pool, alerts, signs_stops_config,
+      # currently, the right number is 6: HTTP pool, alerts,
       # screenplay_config, GTFS, pipeline, health
       actual = children([])
 
-      assert length(actual) == 7
+      assert length(actual) == 6
     end
   end
 end
